+++ Sources/Game_PathFinding.hpp
@@ -61,7 +61,7 @@ namespace Game_PathFinding
 	inline void GenerateFlattenedMap(std::vector<float>& Map, const std::int_fast32_t Width, const std::int_fast32_t Height)
 	{
 		Map.clear();
-		Map.resize(Width * Height, FLT_MAX);
+		Map.resize(static_cast<size_t>(Width) * static_cast<size_t>(Height), FLT_MAX);
 
 		for (std::int_fast32_t x{}; x < Width; ++x)
 		{
